// Frontend/src/lib/protos/database.auth.ts
export enum ProtoProviderType {
    /**
     * @generated from protobuf enum value: oidc = 0;
     */
    oidc = 0,
    /**
     * @generated from protobuf enum value: oauth = 1;
     */
    oauth = 1,
    /**
     * @generated from protobuf enum value: email = 2;
     */
    email = 2,
    /**
     * @generated from protobuf enum value: credentials = 3;
     */
    credentials = 3
}